+++ CHANGELOG.rst
@@ -1,6 +1,12 @@
 Changelog
 #########
 
+4.0.0
+*****
+
+- changed most class members from protected to private
+- cs fixes, added codestyle checks
+
 
 3.0.0
 *****